Rename _to_nerf -> _to_instant_ngp, _to_gaussian_splatting -> _to_nerf
